Review fix + added docs
