+++ docs/index.rst
@@ -54,6 +54,7 @@ And the best thing is you can extend it yourself or... tell us to do it for you 
    api.utils <user_guide/api/utils>
    hpo.utils <user_guide/hpo/utils>
    bots.telegram_bot <user_guide/bots/telegram_bot>
+   logging.chart <user_guide/logging/chart>
    monitoring.lightgbm <user_guide/monitoring/lightgbm>
    monitoring.xgboost <user_guide/monitoring/xgboost>
    monitoring.fastai <user_guide/monitoring/fastai>
