+++ docs/user_guide/logging/chart.rst
@@ -1,0 +1,6 @@
+Chart
+===========
+
+.. automodule:: neptunecontrib.logging.chart
+    :members:
+    :show-inheritance:
